(feat) #43 E2Es now work for Angular 2 post beta
