--- e2e/e2e/typings.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="../typings/main.d.ts" />
